Add function to find best image
For probably internal-to-Google reasons, the default image that accompanies
a lot of gmusicapi resources is a 2:1 aspect ratio, like eg.

http://lh3.googleusercontent.com/Vh9JMSZxAUMhbVhjkkuKRszV-pGHH5nVxx-wTSiSP_p6IHvo7ZPgLf7Rf6SA5Z_EMGISa5oRgKY

LMS _never_ wants 2:1, and ends up internally resizing these to 1:1, but
takes the middle of the image, which even in Google Music is often OK, but
not with these default images.

Image::best() takes a gmusicapi object, the names of the singular and
plural image entries from the API response, and a default fallback,
and returns an Image object created with the right URI.